cleaning up code, use consistent naming
