+++ fallbacks/linux/xsel
