+++ fallbacks/win32/paste.exe
